pbs 64 bits
git-svn-id: svn+ssh://gforge/svn/siconos/trunk@2291 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5

+++ Numerics/src/utils/f2c.h
@@ -62,7 +62,11 @@
 #ifndef _F2C_INCLUDE_H
 #define _F2C_INCLUDE_H
 
-typedef long int integer;
+/* long int => int (64 bits) */
+/* typedef long int integer; */
+
+typedef int integer;
+
 typedef char *address;
 typedef short int shortint;
 typedef float f2c_real;
